auth api reorganisation
